backend/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.0
paypal-server-sdk==1.36.0
